Drivers: hv: Export some functions for use by root partition module
hv_get_hypervisor_version(), hv_call_deposit_pages(), and
hv_call_create_vp(), are all needed in-module with CONFIG_MSHV_ROOT=m.

Link: https://lore.kernel.org/r/1741980536-3865-7-git-send-email-nunodasneves@linux.microsoft.com
Message-ID: <1741980536-3865-7-git-send-email-nunodasneves@linux.microsoft.com>
(cherry picked from commit 21050f619720029e89d365b0e1328eafe088dbd8)

+++ include/asm-generic/mshyperv.h
@@ -250,7 +250,7 @@ int hv_common_cpu_init(unsigned int cpu);
 int hv_common_cpu_die(unsigned int cpu);
 void hv_identify_partition_type(void);
 
-int hv_call_deposit_pages(int node, u64 partition_id, u32 num_pages);
+int hv_call_deposit_pages(int node, u64 partition_id, u64 num_pages);
 
 /**
  * hv_cpu_number_to_vp_number() - Map CPU to VP.
